pyproject.toml
[project]
name = "url-reputation"
version = "1.4.1"
description = "Multi-source URL/domain security reputation checker"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "newuni"}
]
keywords = ["security", "url", "reputation", "malware", "phishing", "virustotal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
full = [
    "dnspython>=2.0.0",
    "python-whois>=0.8.0",
]

dev = [
    "pytest>=8.0.0",
    "ruff>=0.9.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py39"
# The legacy `scripts/` folder is not part of the library surface.
exclude = ["scripts/**"]

[tool.ruff.lint]
# Keep this set tight; we can expand over time.
select = ["E", "F", "I", "B"]
# Initial adoption: ignore a few noisy / legacy patterns; we will tighten later.
ignore = [
  "E501",  # line length
  "E722",  # bare except (legacy code; fix gradually)
  "B904",  # raise ... from err (fix gradually)
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = false
warn_unused_ignores = true
# Initial adoption: do not fail on legacy implicit-optional patterns.
no_implicit_optional = false
check_untyped_defs = false
ignore_missing_imports = true
# Initial adoption: treat mypy as informational; tighten over time.
ignore_errors = true

[project.urls]
Homepage = "https://github.com/newuni/url-reputation"
Repository = "https://github.com/newuni/url-reputation"

[project.entry-points."url_reputation.providers"]
# External packages can register providers here.
# Example: mypkg = "mypkg.providers:provider"

[project.entry-points."url_reputation.enrichers"]
# External packages can register enrichers here.

[project.scripts]
url-reputation = "url_reputation.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["url_reputation"]
